requirements.txt
# Core dependencies
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78

# OCR
pytesseract==0.3.10

# Deep Learning
torch==2.1.1
torchvision==0.16.1

# Transformers & Embeddings
transformers==4.35.2
sentence-transformers==2.2.2
clip-by-openai==0.1.0.post1

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2

# Web Interface
streamlit==1.29.0

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
